Westinghouse WGen9500DF transfer switch interface.
I don't know if anybody else finds this useful? but if you do, please feel free to use this anyway you like.
I have included the schematic pdf file as well as the Kicad.sch file.
I build this on a bread board so I didn't bother with making a PCB for a one-off project, but that would be easy to do 
if anyone feels the urge to do so.